Text transform can set "" on null cells and vice versa.
Previously if a transformation expression resulted in "" or null and the
cell's previous value was "" or null then nothing would be changed.

Partial fix for #820 ( 1 on his list) that opens smallest possible can
of worms re blank/null handling. May break backwards compatibility on
projects that relied on old behaviour, though I think it unlikely anyone
has such a thing.