Support log for Windows folks running refine.bat
Too often we need to ask multiple questions from folks running OpenRefine on Windows.  Let's make this easier for us from a support perspective and just ask the users to paste the contents of the support.log file to us.  We can expand on any support.log enhancements later as we come across the need to add more.