make OpenRefine more robust to misconfigured charset strings,
there are a lot of services out there that are not programmed/configured correctly and that deliver character encoding strings with additional spaces or an semicolon at the end. Regular browsers are robust and don't care, but OpenRefine fails in those cases. This change removes additional spaces, quotation marks and semicolon at the beginning/end of the charset string from the Content-Type HTTP header. Calls to external services that are wrongly configured now return the desired content instead of "HTTP error 200 : OK | ".