Total number of operations for Undo / Redo
Currently, if you revert to the first operation in your project history, you will see Undo / Redo 0. Opening an old project this morning I was confused and thought it was a new one because of this, but in fact I already had many operations on this project, it was just that before closing it I had viewed the file at first operation.

This PR proposes to change the display to Undo / Redo 1/n with n the total number of operations available.