Add menu for searching and replacing
Add a new menu "Replace" in "Edit Cells" > "Common Transforms" and a new interaction box for entering parameters.
Under the hood this is a wrapper around GREL replace() function.
The behavior is the same as replace() except the automatic backslah escaping of "Replace with" parameter.
The 2 main parameters in the box are 2 textaras : "Find" and "Replace with"
2 cumulative options are available for "Find" parameter : "Case insensitive" and "Regular Expression"
1 option is available for "Replace with" parameter : "Do not escape backslash automatically"
A validity control of the pattern is made when "Regular expression" is checked.