Option for matching whole words
Replace Textareas with text Inputs and add labels.
Add an option for matching whole words
Change the behavior of escape option for replacement text:
if checked, \n and \t have a special meaning (newline and tab),
and \\n and \\t must be used for "backslash and n" and "backslash and t"
but other character are escaped : \ means \, \\ means \\, \r means \r etc.