Update the check_running() function in refine
In strict shell environments, the "$?" is reset when evaluated.
Here, the "boolean if" statements compares it to "7", then reset it and compares it to "22".
If the "$?" is intially "22", it is not evaluated and the NOT_RUNNING variable is ignored while it shouldn't. this leads to the error
"Something is already running on http://127.0.0.1:3333/"

To avoid this, the solution is to first affect the value of "$?" to a non-system variable and then evaluate this new variable.