Use release convention for Maven compiler config
Can we move to using latest OpenJDK ?  (I'm using OpenJDK 13)
Regardless, we should use the newer "release" convention instead of source/target which wrecks havok for me on Windows dev environment and switching Maven toolchains for other software.  Maybe Linux users are not prone to env issues as much (a reason Windows created WSL in the first place).  Anyways, some further context on using the "release" flag: https://stackoverflow.com/questions/43102787/what-is-the-release-flag-in-the-java-9-compiler

It's also possible to combine both properties and profile, and just use activation as shown: https://stackoverflow.com/questions/55404641/configure-maven-compiler-plugin-so-it-works-both-with-jdk-8-and-jdk-12-in-sprin