remove unneeded Freebase AGENT_ID
In the past, Freebase editors used Google Refine for making edits to its database and the internal identifier was "/en/google_refine" which equated to a Software Application type with attached metadata and also had ownership privileges for certain Freebase Apps.  Since Freebase is no longer around, this identifier, only used by Freebase, can now be removed.  (This is not a User-Agent header string but was an internal identifier for the Freebase database which no longer exists)