Add recon API key to UI
User is asked for API key when adding a new reconciliation service.
If service requires a key and one isn't provided, it's an error.
Key is stored with the rest of the recon service info (unencrypted).
API key parameter name is retrieved from the manifest and used
together with the provided key value for Guess Types and Reconcile commands.