Fix Google OAuth redirect_uri behind an HTTPs reverse proxy
If OpenRefine runs behind an HTTPs reverse proxy, Jetty 6 doesn't set the
correct client's schema on the HttpServletRequest object, causing the
generated `redirect_uri` for OAuth to always be plain HTTP - which Google
forbids for external OAuth applications.

This commits hardcodes the check, effectively allowing the OAuth flow to
work.

Jetty 7 and newer support a correct way to do this - but Jetty 6 doesn't.