Initial support for schema templates.
Closes #5043.

This adds the ability to start a schema from a previously saved schema template.
A schema template is a schema with some missing values.
It is also possible to pre-define some templates in the Wikibase manifest, so that
those are added to OpenRefine when the manifest is first loaded.

I think it would be useful to add an interface to list existing schema templates, which would enable:
* deleting schema templates
* exporting a schema template to JSON (useful to add it to the manifest afterwards)