Improve error handling in Apply operations dialog.
This is a first step towards #5539 and #5540.

When an operation cannot be executed because its JSON representation is invalid or
it does not apply to the given project, the corresponding error is reported to the user
and the following operations are not executed.

This is not meant to be a final UX, I plan to do many other improvements of this dialog.
This first behaviour is just to make sure we can reuse the logic behind the dialog to apply
operations everywhere else in the tool, for #5539.