Refactor where project row count is updated - fixes #1380, #5387
Fixes #1380 Fixes #5387 probably fixes #3475

Move maintenance of row count in ProjectMetadata from
GetRows (ick!) to the History class where changes are applied.

Update metadata for legacy projects which were created before
row count was added, but do it in a way which doesn't change
the Modified date. (#1380)

Set the row count on project registration instead of waiting
until the first time it's opened (#5387)

Adjust HistoryTest (which seems to be testing the internal
implementation more than the external behavior) to work with
the new implementation.