Migrate CSV/TSV importer to use CloseableIterable
This changes the importing architecture so that importers are passed
not a single InputStream or Reader, but a Supplier of those, so that
they can re-create more as they need. This makes it possible for them
to create a CloseableIterable<Row> for the grids they read.
This move will allow most importers to read their contents lazily from
the original file, avoiding memory bottlenecks.