Add Windows installer using InnoSetup 6
This adds an InnoSetup script and modifies the GitHib Actions workflow to run the necessary steps that produce a self-contained installer file from the contents of the ZIP file of the **Windows (with Java)** build artifact.

So far this is untested on the GitHub Actions CI, but a proper installer is produced like this, when the InnoSetup Command Line Compiler is invoked manually with `"%ProgramFiles(x86)%\Inno Setup 6\iscc.exe" -DMyAppVersion="3.8-SNAPSHOT" -DMyProgramFiles="target\openrefine-3.8-SNAPSHOT" "packaging\openrefine.iss"` from the repository root after manually copying the contents of the folder in the ZIP file to `packaging\target\openrefine-3.8-SNAPSHOT` (that's what I got from running `mvn -B deploy -DskipTests -Dgpg.skip`).