Switch to Apache commons-csv - Fixes #2268
Switch from ancient private fork of opencsv with our
multi-character delimiter patch to the Apache commons-csv
module with supports string delimiters.

There are two types of test failures, one of which I think
is definitely a bug and the other is arguable.

Test changes will go in the next commit to keep them
separate.