Solve Excel importing from URL problem Fixes #6418 (#6537)
- Add a getFileName method in the ImportingUtilities so it can be used to
get the name of the file from the fileRecord and add a test for it.

- Add a new variable (fileName) in the ImportingParserBase method,that
will hold the return value of the new getFileName method of the
ImportingUtilities. So it can be parsed to the parseOneFile method
instead of the fileSource.

This is done to keep the provenance information from being lost.