Make sure that the file page exists before uploading SDC
Checks if the file page has been created after a file has been
uploaded. If not it will wait for a bit and retry. This will
continue until the page exists or the user interrupts, with the
delay between checks increasing up to a maximum. At the moment the
wait time starts at one second and increases by one second per wait.
The max wait time is one minute.

Adds logging for:
* When a file will be uploaded in chunks
* When a chunk is uploaded
* When all chunks have been uploaded
* When the commit request has been made
* Waiting for file page to exist