export: Fix XSS vulnerability
Mitigates CVE-2024-47880.
See https://github.com/OpenRefine/OpenRefine/security/advisories/GHSA-79jv-5226-783f

By setting the contentType parameter to text/html and choosing export
parameters which include a script in the text output, an attacker would
be able to execute arbitrary code in the browser, inside OpenRefine's
origin.

The contentType parameter is not actually supplied in any internal
calls, meaning that the content type declared by the exporter is always
used, so we drop support for this parameter.