Allow uploading new file versions
If there's a match for a file an upload request will be made. The response should contain warnings since the file would already exist. If the only warnings are "exists" and "duplicateversions" another request will be made with `ignorewarnings=1`. This will upload the file as a new version. Allowing these warnings means that you can upload a file with the same name as one that already exists even if the new file is that same as and older version.

An issue is shown when a new version is about to be uploaded to prevent this happening by mistake.

Fixes #5959