Antonin leaving the advisory committee (#7020)
As mentioned [on the forum](https://forum.openrefine.org/t/transitioning-out-of-the-project/1879), I am in the process of reducing my involvement in OpenRefine. Leaving the advisory committee is one step of that process. I intend to continue development and other activities until the end of March, but I don't need to be in the committee for that.

Perhaps it would be worth thinking about onboarding new members on the advisory committee.