fixes #7207
This change addresses how the "last" pagination navigation button sends
the ending point. The current behavior is that it sends the index of the
last row relative to a faceted search, and this change is to send the
index of the absolute last row in the project.
This also includes a change to omit the `nextPageStart` variable if
there are fewer than `limit` elements in the response. This will cause
the `next` and `last` pagination buttons to be marked as inactive, as
there are no more elements to paginate through.