Send absolute row index in request when "last" pagination button is clicked (#7211)
* fixes #7207

This change addresses how the "last" pagination navigation button sends
the ending point. The current behavior is that it sends the index of the
last row relative to a faceted search, and this change is to send the
index of the absolute last row in the project.
This also includes a change to omit the `nextPageStart` variable if
there are fewer than `limit` elements in the response. This will cause
the `next` and `last` pagination buttons to be marked as inactive, as
there are no more elements to paginate through.

* set totalRows to project row length and add tests

This commit changes the behavior of GetRowsCommand to return the total
number of rows in a project as the `totalRows` field, rather than
returning the filtered row count.
This commit also adds two tests to GetRowsCommandTest, to test that
pagination still works as expected when facets are applied.