Migrate from postCoreProcess to postOperation
Closes #5539.

This changes the way operations are submitted to the backend,
avoiding the definition of one command for each operation. Instead,
operations are submitted to the generic 'apply-operations' command,
using their JSON representation. In the future, the command classes
may be removed, getting rid of an unnecessary additional layer.

See also #5559 for an earlier review in the `delta` branch.