Add value1 and value2 to Jython and Clojure
In UserDefinedDistance.java (#6612) we introduced two new variables
`value1` and `value2`. In GREL these are accessible as "globals".

But in Jython and Clojure we encapsulate the expressions in functions
and only pass certain variables as parameters
(value, cell, cells, row, rowIndex).

This means that in Jython and Clojure expressions we can not access
variables like `columnName`.

This also means that we somehow have to make the new variables
`value1` and `value2` accessible to Jython and Clojure expressions.

Without rewriting the whole Jython and Clojure integrations
it felt most idiomatic to also introduce these variables as parameters
to the Jython and Clojure functions that encapsulate the expressions.

But it is still somewhat unsatisfying that we have a different pool of
available variables in GREL and (possibly) each other language integration.