2024 governance update (#7023)
* Antonin leaving the advisory committee

As mentioned [on the forum](https://forum.openrefine.org/t/transitioning-out-of-the-project/1879), I am in the process of reducing my involvement in OpenRefine. Leaving the advisory committee is one step of that process. I intend to continue development and other activities until the end of March, but I don't need to be in the committee for that.

Perhaps it would be worth thinking about onboarding new members on the advisory committee.

* Revise Governance following discussion https://forum.openrefine.org/t/improving-transparency-advisory-committees-role-and-community-involvement/1672/37 , clarify code of conduct language

* governance typo fix

* Update GOVERNANCE.md

Adjust reference to the Member Election section

* Update GOVERNANCE.md

* Apply suggestions from code review

* move Advisory Committee specific content to the policies section

* narrowed the funding application to apply to the Advisory Committee and Core Dev Group when funds are managed by CS&S

* add intro to the policy section

* remove reference to OpenRefine mission and value in the funding policy

* Update GOVERNANCE.md

How about this as a way of softening the conflicts between the two paragraphs?

* Update GOVERNANCE.md

removes period from item in list to keep in the same format as the other items on the list

* Update GOVERNANCE.md

removes 'your' to make the updated list item flow better with the rest of the list items

* Update GOVERNANCE.md

punctuation

* Apply suggestions from code review

Co-authored-by: Thad Guidry <thadguidry@gmail.com>

* Update GOVERNANCE.md

cleaning wording on how we introduce the policies section.

* Update GOVERNANCE.md

moving the election process as part of the governance.

* Update GOVERNANCE.md

Co-authored-by: Thad Guidry <thadguidry@gmail.com>

* Update GOVERNANCE.md

* Apply suggestions from code review

Co-authored-by: Tom Morris <tfmorris@gmail.com>

---------

Co-authored-by: Antonin Delpeuch <antonin@delpeuch.eu>
Co-authored-by: Esther Jackson <64559180+ej2432@users.noreply.github.com>
Co-authored-by: Thad Guidry <thadguidry@gmail.com>
Co-authored-by: Tom Morris <tfmorris@gmail.com>