ExporterManager: add an explicite extension point
Prior to this change one would extend the export menu by pushing directly
to an array. This comes with several issues:

 - The interface isn't actually controlled by use and is increadibly large.

 - We can't document the interface in a structured way.

 - Extensions ability to modify the menu isn't limited.

This change does not break existing interfaces as the underlying array  is
untouched. Some extensions do more than pushing to the array, however, that
could be considered a side effect of the interface rather than an intended
feature. If there are legitime use-cases beyond pushing additional exension
points could be added.