Improve exporter error reporting & XLS/XLSX error handling. Fixes #7368 & #352. (#7459)
This includes a fix for the specific XLS row limit exceeded error (#7368) as well as better error reporting for exporters in general, which we seem to have broken sometime in the recent past. It also fixes #352 (silent truncation of large cells) and provides a better fix for #351 (almost silent truncation of sheets with too many columns).

* Fix exporter error handling

- Add ExporterException & refactor GetRows command to use it instead of specific SQL exporter exception
- Make SQL exception a subclass of new ExporterException
- Fix HTTP response headers for error case, so HTML error page gets rendered again
- remove obsolete CSRF comment

* Throw exception when XLS(X) row/col/cell limits are exceeded

- BEHAVIOR CHANGE - column and cell limits no longer cause silent truncation
- Row limit error handled more cleanly
- Adjust / add unit tests for new behavior
- Add e2e tests for Excel limit errors