Remove unused options Properties from createProcess API - BREAKING
This is unused internally, always passed as an empty Properties object
by the caller and always ignored by the callee. The old method has been
removed outright, any overriders or callers will need to be updated.

If we ever needed it for some reason, we'd probably introduce it as as
Map<String, String> or Map<String, Object>